Move helpers to tests folder #20 #82
